requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
supabase>=2.3.0
neo4j==5.14.1

# LLM & Embeddings
# openai>=1.6.1,<2.0.0
# anthropic>=0.7.7
# langchain>=0.0.350
# langchain-openai>=0.0.2
# langchain-anthropic>=0.0.1
# sentence-transformers>=2.2.2
# LLM & Embeddings (고정 추천)
openai>=1.10.0
anthropic>=0.34.0
langchain==0.1.6
langchain-openai==0.0.6
langchain-anthropic==0.1.0
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0

# HTTP 클라이언트 (의존성 충돌 해결)
# supabase 2.3.0+와 anthropic/openai 호환을 위해 명시적 버전 지정
httpx>=0.25.0,<1.0.0

# Text Processing
# tiktoken==0.5.1
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
tqdm==4.66.1
numpy==1.24.3
requests==2.31.0

